Fixed Bug #{1545}
The situation can be reproduced by diconnecting a client (console or ws) while jnem is busy doing analysis.
This is easiest at a jnem tock where the tock analysis takes several seconds such as with the 2ID scenario.

Configure jnem sim/fedsim to use the 2ID scenario and disconnect a client while jnem is "busy".

Verify that the logs now contain messages such as the following rather than the bgerror that was previously being generated.

 101123ZJAN05 detail  wsman  Error sending to <33303>: error flushing "sock26": broken pipe
 101123ZJAN05 detail  wsman  Disconnect: 'console2' at <33303>