Fixed Bug #{2117}
The problem arose because the "lock" and "unlock" methods take not a
parameter name, but a pattern; all parameters whose names match the
pattern are locked or unlocked.  It's not obviously an error if no
parameters match the pattern.

However, it *is* an error, practically speaking, so these operations 
will now throw the following error when the pattern matches no 
parameters:

  Pattern matches no parameters: "<pattern>"