Fixed Bug #{2349}
Done.

I elected not to add isearch automatically to every texteditor(n) widget;
you generally want to enable logging, and that implies having somewhere to
show the messages.  That means that generally you want to do it at a 
higher level in the application.

However, isearch has been added to the modeditor, and to the rotext
widgets in the cmdbrowser and winbrowser, so that now it's generally 
available in the debugger(n) window.