Fixed Bug #{2360}
It is now possible to specify -athresh and -dthresh for the four kinds of
input, populating the athresh and dthresh columns in gram_effects.  Moreover,
these columns are included in gram_sat_effects and gram_coop_effects, and
are displayed in the relevant tabs in mars_gram(1).

While testing that, I found a bug in gramdb(5): the coop0 column was
being defaulted to the rel column's value in gramdb_fg and gramdb_nfg.
Fixed.