Fixed Bug #{2362}
Experiments and discussions with Richard Hipp yesterday revealed that
it's a no-no to have the body of an [rdb eval] update a table that it's
iterating over.  The results are (to quote Richard) "undefined".  Thus,
although it's been working for us, we need to change our code so as not
to get bitten by some future change to SQLite3.  

I tried two alternate approaches in the routine

   ComputeNominalContributionsForSlopeEffects

1) Wrapping the [rdb eval] in a [foreach]
2) Caching the updates in a list and doing them after the 
   [rdb eval] returns.  

Neither approach changed the performance significantly for my test case.
I chose #2 for this routine, as it results in slighly prettier code.