Fixed Bug #{2362}
The previous commits left a situation where a slope effect's "acontrib"
can be non-zero even though the effect contributed nothing during the
time step.  If a slope goes temporary to zero, then in time steps when
it is zero the "acontrib" will be neither written nor read, and will 
therefore contain the previous non-zero value.  This is not an error,
but it *is* potentially misleading.  

This commit arranges for acontrib to be zeroed when ncontrib is set, so
that this misleading situation does not arise.