Fixed Bug #{2459}
After more reflection, I decided that we needed a more general <<Context>>
event.  Any right-click anywhere on the chart (except in a popped-up legend
window) will result in a <<Context>> event.  The %d substitution indicates
what kind of thing was clicked on.  See the man page for details.

This commit includes the code changes, an update to the test script, and
documentation for the <<Commit>> event in the man page.