Fixed Bug #{2709}
Fixed. There were two issues.

1. At the start of the reload, the browser's contents was cleared 
using the public "clear" method, which calls -selectioncmd as a 
matter of course.  I added a helper method, called by both 
ReloadContent and by "clear", that just does the clear without
calling -selectioncmd.

2. At the end of the reload, the browser's content is sorted by
SortData, which is usually called because of a user interaction
and hences calls -selectioncmd.  I added "SortDataNotify" for the
user interaction case, so that SortData need not call 
-selectioncmd.

At present, reloading the data will always call -selectioncmd
once.  If the selected entities are still present, the new 
selection will be the same as the old; that's OK.  The erroneous
transients have been removed.