Fixed Bug #{2711}
I made three changes to the capability as described.

1. Many routines of this type are implemented by sqlib(n); 
sqldocument(n) simply delegates to them.  I followed this
model.

2. As specified, the "grab" result includes the column names.
I wanted to include these so that "ungrab", which needs the
column names, could be used with views as well as tables, 
at least in theory.  But it turns out that "PRAGMA table_info()"
will return the column names for views as well; thus, the
"names" data was unnecessary. "grab" now returns 

   <table1> <values1> <table2> <values2> ...

3. The "ungrab" command expects to be called like this:

   $db ungrab {*}$data

where

   set data [$db grab ...]

This makes it more useful for populating databases during testing.