Fixed Bug #{2717}
Done.  

In addition to checking the calling syntax for "ungrab", I changed the
format returned by "grab" and "grabbing_delete" and expected by "ungrab".
The column values are now SQL-quoted, so that they can be formatted
directly into SQL queries without using variable references; this allows
these routines to grab and restore NULL values properly.

Because grabbing a data set might have other uses in a Tcl application,
the "grab" method now supports a "-tcl" option which grabs the data 
without SQL-quoting, so that it can be used directly.