Fixed Bug #{2830}
Whoops!  sqldocument(n) uses its instance name as the subject
for the notifier events sent by "monitor script" and
"monitor transaction".  Athena expects the subject to be the
name of the scenariodb(n) wrapper.  Added a -subject option, so
that the client can tell the instance what subject name to use.