Fixed Bug #{2850}
Revised the test_newkeyfield.tcl test script.  Instead of using
test.db as a source of data, it copies and modifies the schema
from test_orderdialog.tcl.  It now has an integer grid table, 
"igrid", and a text grid table, "tgrid", with associated indices.

This will now be the standard test script for this module.