Fixed Bug #{2916}
The computational changes have been made, including replacing the 
Z.fi factor in the numerator and denominator of the general case with
max(Z.fi,D.fgi).  In the simple test cases run by the test suite, the
results are the same.  We need to look at the results in practice, and
that means updating Athena to allow editing of gamma and the entity
commonality coefficients.