Fixed Bug #{3011}
Done!

No I&T is necessary, this module is simply there in case Mars is built on a non-Linux platform (ie. Mac).

The Tcl binding API has changed a bit, however; the creation of a gtifreader(n) and the reading of a GeoTIFF file are separated to make error tracking easier.