Fixed Bug #{3122}
Fixed.

For "messagebox gets" and "messagebox pick", pressing the window manager's
"x" button should be the same as pressing "Cancel"; this is now the case.

For "messagebox popup", the case is more complicated.  The dialog can have
any number of buttons, none of which need be a "cancel" button.  Consequently,
I've added an -onclose option, which names the button symbol to return when
the "x" button is pushed.  This can be set to the symbolic constant for the
"cancel" button (or to anything else the user prefers).  Since "Cancel" is
usually the leftmost (first) button, the default for -onclose is the first
button in -buttons.