Fixed Bug #{3232}
The modeditor's behavior when calling the format command callback is modified to append the modeditor's window name to the callback.

This is so users of the modeditor can use it if need be, such as when a messagebox needs to be popped up over it.

Updated the docs as well.

The modeditor is a developer tool, as such, no specific I&T