Fixed Bug #{3236}
Done!

The class method now has an optional argument called "-norefresh" that, when specified, triggers this behavior.
Without this option objects are refreshed like they always have.

This is put in for use by the game truth sent for the webserver in the JNEM console.
There is a lot of data, so refreshing it just isn't practical, but it's nice to still use the normal update methods in the gt server/client interface to get data from the sim to the console.