Fixed Bug #{3446}
OK, now we simply don't have FK constraints at all in the uram_civrel_t,
uram_frcrel_t, and ucurve_contribs_t tables.    It was necessary to
make ucurve(n)'s "clear" method purge the ucurve_contribs_t table, as with
the previous "solution".  The problem is that FK constraints are checked after
deletions even with "ON DELETE NO ACTION"; and since clearing the database
involves clearing tables in no particular order, NOT using 
"ON DELETE CASCADE" on all FK constraints leads to constraint failures.

This time I made the changes in the Athena context, and tested the effect on
Athena before committing.