Fixed Bug #{3468}
Having committed the "resonance" code, I discovered that the term "resonance" 
had been reserved for a different variable in the Info Ops model, one based on
what I was calling "resonance" but distinct from it.  After some discussion
with RGC, we decided that what mam(n) is computing in this case is the 
"congruence" between the hook and the entity's belief system.  Thus, the 
"resonance" of a semantic hook with a civilian group will depend on the 
"congruence" of the hook with the group's belief system.

So this commit changes the word "resonance" to "congruence" throughout the
mam(n) code, tests, and documentation.