Fixed Bug #{3488}
Fixed!

Turns out this problem was encountered because the ucurve logic was updating an SQL table with columns that have a constrained interdependency.

The sum of two columns cannot exceed 1.0, but ucurve was updating one value at a time, so although the sum of the new values would not be > 1.0 the constraint is violated.

The fix is to update the values in one SQL update rather than multiple.

The test suite is also updated because error reporting now provides more information when the constraint is violated.

Retest.