Fixed Bug #{3623}
Done!

There is a -places option that expects an integer for the number of decimal places to display with the comma formatted number.
If the number of places requested is longer than needed then zeroes are padded.
If it's shorter then rounding takes place.

The man page is updated as well.