Fixed Bug #{3624}
Done!

Not much was necessary other than to delegate the option to the underlying TkTable, now users of cmsheet(n) objects can react to a user moving around in the cells of the sheet.

The docs are updated, this functionality will be tested as part of the 6x6 econ model testing.