Fixed Bug #{3637}
Done!

The interface to cellmodel(n) also has an 'initial' method that returns the cell models initial state prior to the last attempt to solve.

The method returns a dictionary of fully qualified cell names and values. This data is useful for debugging purposes.

The man pages are also updated.