Fixed Bug #{3713}
Added lazy updating to reportbrowser(n), including a -reloadon option.  This
implementation isn't as clean as it could be; the widget supports a couple of
different ways of being poked to update itself, and since we plan to stop 
using it in the mid-term it's not worth cleaning all of that up.