Fixed Bug #{3729}
Done!

Mars now contains a binding that uses the libTiff and libGTiff libraries that are built from the OSGeo GeoTIFF extension of the basic TIFF library.

No specific tests are necessary, this is just the binding. Athena will still need to be modified to import and use GeoTIFF map images.