Fixed Bug #{3735}
Done!

The geotiff(n) read command will now return a dictionary of pertinent data from a GeoTIFF file for an image that is geo-referenced using the GeoTIFF model type GEOGRAPHIC. There are other model types, but this is the one that we are likely to have to deal with. The reader may need to be augmented if that is not the case.

No specific I&T, it will be tested once Athena accepts the import of GeoTIFF images and geo-references them.

The geotiff(n) manpage is updated.