Fixed Bug #{3902}
Done.

The change is transparent to an end user since they use order dialogs which were filling in the defaults in thier windows.

This change is for clients that use the order 'send' method directly.

It is sufficient to observe that all order processing works normally during the course of testing.